+++ src/__init__.py
@@ -3,8 +3,9 @@ __version__ = '.'.join(__version_info__)
 
 from .wrappers import (ObjectProxy, CallableObjectProxy, FunctionWrapper,
         BoundFunctionWrapper, WeakFunctionProxy, resolve_path, apply_patch,
-        wrap_object, function_wrapper, wrap_function_wrapper,
-        patch_function_wrapper, transient_function_wrapper)
+        wrap_object, wrap_object_attribute, function_wrapper,
+        wrap_function_wrapper, patch_function_wrapper,
+        transient_function_wrapper)
 
 from .decorators import decorator, synchronized